requirements.txt
pyyaml>=5.3
python-dotenv>=0.10.3